🚚 [Move]: value_objects/ 中間階層を剥がし VO ファイルをドメイン直下へ配置
DDD のフォルダ命名で「Value Object」はパターン分類であり、ドメイン階層に
持ち込むと domain-oriented にならないため、config/project/task 各ドメイン
配下の `value_objects/` フォルダを廃止し、子ファイルをドメイン直下へ昇格。
VO であることは型定義（private フィールド + `try_from_*`）で表現する方針。

- config/value_objects/column_name{,_tests}.rs → config/
- project/value_objects/project_root{,_tests}.rs → project/
- task/value_objects/{label,task_file_name,task_file_path,task_title}{,_tests}.rs → task/
- 親 *.rs の `pub mod value_objects;` を子モジュール直接列挙へ
- `use crate::*::value_objects::*` 参照を全箇所更新
