--- src-tauri/src/config/value_objects.rs
@@ -1,1 +1,0 @@
-pub mod column_name;
