+++ src-tauri/src/project.rs
@@ -1,2 +1,2 @@
 pub mod open;
-pub mod value_objects;
+pub mod project_root;
