--- src-tauri/src/project/value_objects.rs
@@ -1,1 +1,0 @@
-pub mod project_root;
