+++ src-tauri/src/task/path_normalization.rs
@@ -1,6 +1,6 @@
 //! Task の file path 文字列を正規化するための共通 helper。
 //!
-//! `task::index` と各 VO（`task::value_objects::task_file_path` 等）の双方から
+//! `task::index` と各 VO（`task::task_file_path` 等）の双方から
 //! 参照される pure helper を集める。VO が Aggregate (`task::index`) に依存する
 //! 責務逆転を避けるため、独立モジュールとして切り出している。
