--- src-tauri/src/task/value_objects.rs
@@ -1,4 +1,0 @@
-pub mod label;
-pub mod task_file_name;
-pub mod task_file_path;
-pub mod task_title;
